Restructure: make user dialog hook
