+++ src/dashboard/src/Layout/AppBar/UserButton.tsx
@@ -8,13 +8,6 @@ import {
 
 import {
   Button,
-  Dialog,
-  DialogTitle,
-  Divider,
-  Link,
-  List,
-  ListItem,
-  ListItemText,
   makeStyles,
   createStyles
 } from '@material-ui/core'
@@ -23,7 +16,7 @@ import {
 } from '@material-ui/icons'
 
 import UserContext from '../../contexts/User'
-import TeamContext from '../../contexts/Team'
+import useUserDialog from '../../hooks/useUserDialog'
 
 const useStyles = makeStyles(() => createStyles({
   'root': {
@@ -32,59 +25,19 @@ const useStyles = makeStyles(() => createStyles({
 }))
 
 const UserButton: FunctionComponent = () => {
-  const { email, password, givenName, familyName } = useContext(UserContext)
-  const { currentTeamId } = useContext(TeamContext)
-  const api = window.location.origin + `/api/teams/${currentTeamId}/jobs` +
-    `?email=${encodeURIComponent(email || '')}&password=${encodeURIComponent(password || '')}`
-  const [open, setOpen] = useState(false)
-  const handleClick = useCallback(() => {
-    setOpen(true)
-  }, [])
-  const handleClose = useCallback(() => {
-    setOpen(false)
-  }, [])
+  const { givenName, familyName } = useContext(UserContext)
+  const { open } = useUserDialog()
   const styles = useStyles()
   return (
-    <>
-      <Button
-        variant="outlined"
-        color="inherit"
-        classes={styles}
-        startIcon={<AccountBox/>}
-        onClick={handleClick}
-      >
-        {`${givenName} ${familyName}`}
-      </Button>
-      <Dialog
-        maxWidth={false}
-        open={open}
-        onClose={handleClose}
-      >
-        <DialogTitle>User Account</DialogTitle>
-        <Divider/>
-        <List dense disablePadding>
-          <ListItem>
-            <ListItemText primary="Email" secondary={email}/>
-          </ListItem>
-          <ListItem>
-            <ListItemText primary="Password" secondary={password}/>
-          </ListItem>
-          <Divider/>
-          <ListItem>
-            <ListItemText
-              primary="Try an API"
-              secondary={api}
-              secondaryTypographyProps={{
-                component: Link,
-                href: api,
-                target: '_blank',
-                rel: 'noopener noreferrer'
-              }}
-            />
-          </ListItem>
-        </List>
-      </Dialog>
-    </>
+    <Button
+      variant="outlined"
+      color="inherit"
+      classes={styles}
+      startIcon={<AccountBox/>}
+      onClick={open}
+    >
+      {`${givenName} ${familyName}`}
+    </Button>
   )
 }
